requirements.txt
fastapi
gitpython
pydantic
graphviz